feat: enables the farmOS inventory module (#16)
FarmData2 uses the farmOS inventory module to track the inventory of
assets (e.g. the number of trays available for transplant, or the number
of row/feet available for harvest, etc). To do so, the inventory module
must be enabled. This module is not enabled by default, so this PR
enables it when the sample database is built.