fix: removes ssl key creation from sample DB build (#26)
The build of the sample database had included the generation of SSL keys
for the API. However, those keys are not stored in the database but
instead in the farmos container file system. Thus, generating them in
the sample database did not have the intended effect. They are not
generated during the build of the farmos docker container.

See: https://github.com/FarmData2/FD2-Images/pull/4