fix: clarifies and adjusts the FarmData2 log categories (#38)
Currently defines 3 FarmData2 specific log categories:
```
seeding_cover_crop,For logs associated with the seeding of a cover crop.
seeding_direct,For logs associated with direct seedings.
seeding_tray,For logs associated with seedings in trays.
```

These are attached to any relevant logs. For example, an activity (e.g.
a `tillage`) would also have a category of `seeding_direct` if it is
associated with a direct seeding event.