chore: copies instead of links FD2 libraries (#44)
The prior script `linkFD2Libs.bash` created sym links to the libraries
in the FarmData2 repository. This worked, but required changes to the
imports in the `lib.js` files in the `farm_fd2` modules (e.g.
`tray_seeding`). Rather than make those changes, which would seem out of
place in FarmData2, the `linkFd2Libs.bash` script was replaced by the
`buildFD2libs.bash` script. This script make a copy of the relevant
`.js` files and edits the `import` statements such that they work in
this repo.