fix: updates tray seeding data to allow better testing (#46)
Tray seeding data was split into two sets, one that will not be
transplanted and one that will be transplanted (later). Only the tray
seedlings that will not be transplanted have been added at this point.
Some of those tray seedings were created with 0 trays to allow better
testing of the filtering of the seedlings api result.